requirements.txt
tweepy
python-dotenv
pyspark
pandas
kafka-python
psycopg2-binary
